Move DCR restart test back into runner package
Drops the package authserver_test workaround introduced to break the
runner -> authserver import cycle that blocked extending
pkg/authserver/integration_test.go. The test's actual subject is the
runner-side DCR-store wiring (it goes through runner.NewEmbeddedAuthServer
and asserts on runner.EmbeddedAuthServer.DCRStore()), so the runner
package is the natural home and matches the location of its sibling
TestNewEmbeddedAuthServer_DCRBoot / _ClosesStorageOnError tests.

The runner-package newMockAuthorizationServer helper replaces the
duplicate newMockUpstreamAS the cross-package placement forced; the
"DO NOT COPY THIS A THIRD TIME" tripwire on it is dropped now that the
duplication is gone.
